자료실 파일 첨부기능 구현

+++ semiprj/src/main/java/com/kh/semiprj/dto/PdsDto.java
@@ -9,7 +9,7 @@ import lombok.Data;
 
 @Data
 public class PdsDto {
-	private long pdsNo;
+	private int pdsNo;
 	private String pdsWriter;
 	private String pdsTitle;
 	private long pdsReadcount;
